# docs/api/openapi.yaml
swagger: '2.0'
info:
  title: netobs unified API
  version: 2.5.6
paths:
...
  /api/v1/memory:
    get:
      description: pod별 cAdvisor 종류별 메모리(working_set / rss / cache / swap bytes)와
        limit, OOM 위험(working_set/limit), 지배 종류를 돌려준다. rss(anonymous)는 OOM을 유발하고 cache는
        reclaimable이라, working_set이 rss로 채워졌는지 cache로 채워졌는지 구분해 실제 압박 여부를 판정한다. swap은
        노드 swap이 켜진 경우만 채워진다. ?namespace로 필터한다.
      produces:
      - application/json
      tags:
      - interference
      summary: 메모리 병목 분해
      parameters:
      - type: string
        description: namespace 필터 (생략 시 전체)
        name: namespace
        in: query
      - type: integer
        description: 상위 N pod (1-200, 기본 30)
        name: limit
        in: query
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.MemoryResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
  /api/v1/node-vitals:
    get:
      description: '노드의 CPU 와 memory 사용률 퍼센트 (노드 내 pod 합산 사용량을 노드 allocatable 로 나눈
        점유율, #313), GPU 사용률과 GPU 메모리를 instant query 로 읽어 노드 단위로 돌려준다. cadvisor 와 gpuobs
        원시 게이지 기반이며 압박 score 로 변환하지 않는다. limit 근접도 의미는 pressure 계열이 담당한다. 프론트가 주기
        폴링으로 실시간 값을 받는 경로이며, 시계열 range 는 trends 가 담당한다.'
      produces:
      - application/json
      tags:
      - interference
      summary: 노드 raw 사용률 (Vitals)
      parameters:
      - type: string
        description: 대상 노드 (DNS-1123 형식)
        name: node
        in: query
        required: true
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodeVitalsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
  /api/v1/node/{node}:
    get:
      description: '노드의 4 차원 pressure 와 health, 종합(overall), dominant 차원과 신뢰도, 그 노드
        위 top 압박 pod 를 모아 한 노드의 상태를 한 응답으로 돌려준다. health 는 node:*_health_score:5m 룰(cluster
        health 의 노드 차원 판)이고, 신뢰도는 압박 top1 과 top2 차원 격차라 gpu-rca 와 동일 축이다. status 는
        차원별 pressure 등급(전 차원 worst, memory 는 node_exporter 실측 사용률이라 일반 임계 대신 usage
        임계 0.85/0.95 로 환산해 health 불감대·usage 축 설계와 정합)과 노드 사용량(CPU/memory 점유율, allocatable
        분모, 0.85 warn/0.95 degraded) 등급과 차원별 health 최솟값 등급과 이 노드를 가리키는 firing alert(severity
        critical 은 degraded, 그 외 warn) 등급의 worst-of 합성(#324, #325)이며, status_basis
        가 결정 신호(pressure/usage/health/alert)를 담는다. GPU 사용률은 포화가 정상 활용일 수 있어 등급 입력에서
        제외한다. 등급 어휘 ok/warn/degraded 는 overview 와 node-map 의 healthy/warning 과 같은
        입력 신호(pressure, firing alert)를 공유하고, down(ready 기반) 판정은 node-map 소관이다.'
      produces:
      - application/json
      tags:
      - interference
      summary: 노드 1대 전체 압박 상황
      parameters:
      - type: string
        description: 노드 이름
        name: node
        in: path
        required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodeResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
  /api/v1/node/{node}/pods:
    get:
      description: '노드에 스케줄된 pod 별로 신원 (namespace 와 pod 와 uid) 과 CPU (limit 대비 percent
        와 절대량 cores), memory (limit 대비 percent 와 working set bytes), network bytes
        rate (netobs pod bytes 의 5분 rate 합산) 와 상태를 한 응답으로 돌려준다. percent 는 limit 분모라
        limit 없는 pod 는 생략되고 절대량 필드로 표현된다 (#328 과 동일 규약). severity 는 pod pressure score
        3종 최대의 환산 (low/elevated/high) 이고, 종료 pod 와 미관측 사유 (unobserved_reason, no_traffic
        포함 #342) 는 pods API 의 규약 (#314, #320) 을 그대로 쓴다. 관측 판정 소스는 rate 결과 존재라 (pods
        API 는 시리즈 존재) 카운터 샘플이 부족한 신규 pod 의 짧은 창에서는 여기서만 일시적으로 사유가 붙을 수 있고 1분 내 자연
        수렴한다. 정렬은 namespace 와 pod 사전순으로 결정적이며 이상 우선 정렬은 severity 를 근거로 표시 계층이 수행한다.
        at 파라미터로 사건 시점을 재구성할 수 있다.'
      produces:
      - application/json
      tags:
      - interference
      summary: 노드 하위 pod별 자원 사용량 목록
      parameters:
      - type: string
        description: 노드 이름 (DNS-1123 형식)
        name: node
        in: path
        required: true
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodePodsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
  /api/v1/nodes:
    get:
      description: 노드별 이름, uid, 내부/외부 IP, Ready 상태, 역할(control-plane 등), 버전, capacity(cpu/memory/gpu)를
        kube-state-metrics 기반으로 돌려준다. 다른 API의 node 라벨과 동일 키로 매핑한다.
      produces:
      - application/json
      tags:
      - inventory
      summary: 노드 인벤토리
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodesResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
  /api/v1/pod/{namespace}/{pod}:
    get:
      description: 단일 pod 의 기본 정보 (uid 와 node 와 pod_ip 와 owner) 와 차원별 health 3종 (pod
        단위 pressure score 의 1 - score 환산), vitals (limit 대비 CPU 와 memory 사용률과 working
        set, CPU 절대 사용량), cgroup CPU throttle 상세 (throttled 비율과 5분 throttled/total
        periods 와 limit cores, cadvisor 가 노출하는 클러스터에서는 quota 와 period), 네트워크 종합 (pod
        입도 재전송과 drop rate, 최대 smoothed RTT) 을 한 응답으로 합성한다. cpu_percent 는 limit 대비
        비율이고 cpu_usage_cores 는 5분 rate 의 절대량 (cores) 이라 limit 없는 pod 는 cores 만 존재한다
        (#328, memory 의 working set 절대값과 대칭). RTT 는 분위수가 아닌 최대값이라 필드명이 max_srtt_seconds
        다. 미관측 pod 는 필드 생략과 summary 사유로 graceful 처리한다.
      produces:
      - application/json
      tags:
      - interference
      summary: pod 단위 종합 상세
      parameters:
      - type: string
        description: pod 네임스페이스 (DNS-1123 형식)
        name: namespace
        in: path
        required: true
      - type: string
        description: pod 이름 (DNS-1123 형식)
        name: pod
        in: path
        required: true
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.PodDetailResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
  /api/v1/pods:
    get:
      description: '파드별 namespace, 이름, uid, pod IP, host IP, node, workload(created_by),
        priority, phase, qos, 관측 커버리지(observed)를 kube-state-metrics 기반으로 돌려준다. observed
        는 netobs 의 eBPF 시리즈가 존재하는지로, false 면 unobserved_reason 이 미관측 사유를 분류한다 (#320,
        #342): agent_absent 는 노드에 netobs 미배치, host_network 는 pod IP 가 host IP 와 같아
        IP 귀속이 성립하지 않는 상태로 cgroup 힌트가 학습되면 live 로 전환될 수 있는 시점 의존 분류다. no_traffic 은
        agent 의 소켓 스캔이 pod netns 의 무소켓을 증명한 상태 (unix socket 전용 pod 등) 로 시리즈 부재가 정상이고,
        no_data 는 관측 가능한데 시리즈 부재다. 종료 pod (phase Succeeded/Failed) 는 telemetry 부재가
        정상이라 observed=false 에 사유가 생략된다 (#314). ?namespace 로 필터한다. 다른 API의 src_namespace/src_pod/pod_uid와
        동일 키로 매핑한다.'
      produces:
      - application/json
      tags:
      - inventory
      summary: 파드 인벤토리
      parameters:
      - type: string
        description: namespace 필터 (생략 시 전체)
        name: namespace
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.PodsResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
definitions:
...
  internal_correlation_api.MemoryResponse:
    type: object
    properties:
      generated_at:
        type: string
      pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.PodMemory'
      summary:
        type: string
      window:
        type: string
  internal_correlation_api.NodeCapacity:
    type: object
    properties:
      cpu:
        type: number
      gpu:
        type: number
      memory_bytes:
        type: number
...
  internal_correlation_api.NodeInventory:
    type: object
    properties:
      capacity:
        $ref: '#/definitions/internal_correlation_api.NodeCapacity'
      external_ip:
        type: string
      internal_ip:
        type: string
      kernel_version:
        type: string
      kubelet_version:
        type: string
      name:
        type: string
      os_image:
        type: string
      ready:
        type: boolean
      roles:
        description: 'Roles 는 kube_node_role 기반 노드 역할 (control-plane 등) 이다 (#248).
          역할 라벨이 없는

          worker 노드는 생략된다.'
        type: array
        items:
          type: string
      runtime_version:
        type: string
      uid:
        type: string
...
  internal_correlation_api.NodePodPressure:
    type: object
    properties:
      dimension:
        type: string
      pod:
        type: string
      pressure:
        type: number
  internal_correlation_api.NodePodUsage:
    type: object
    properties:
      cpu_percent:
        type: number
      cpu_usage_cores:
        type: number
      memory_percent:
        type: number
      memory_working_set_bytes:
        type: number
      namespace:
        type: string
      network_bytes_per_sec:
        type: number
      phase:
        type: string
      pod:
        type: string
      severity:
        type: string
      uid:
        type: string
      unobserved_reason:
        type: string
  internal_correlation_api.NodePodsResponse:
    type: object
    properties:
      generated_at:
        type: string
      node:
        type: string
      pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.NodePodUsage'
      summary:
        type: string
...
  internal_correlation_api.NodeResponse:
    type: object
    properties:
      confidence:
        description: 'Confidence 는 dominant 차원 판정 신뢰도 (0-1, #264) 다. 압박 top1 과 top2
          차원의 격차로,

          gpu-rca 의 신뢰도와 동일 축이라 한 차원이 지배적일수록 1 에 가깝다.'
        type: number
      dominant_dimension:
        type: string
      generated_at:
        type: string
      health:
        description: 'Health 는 차원별 health score (0-1, #264) 다. node:*_health_score:5m
          룰 기반이며 cluster

          health 와 동일 산식의 노드 차원 판이라 GPU 와 비GPU 노드가 같은 해석을 공유한다.'
        type: object
        additionalProperties:
          type: number
          format: float64
      node:
        type: string
      overall:
        type: number
      pressure:
        type: object
        additionalProperties:
          type: number
          format: float64
      status:
        type: string
      status_basis:
        description: 'StatusBasis 는 status 등급을 결정한 신호다 (#324, #325, pressure / usage
          / health / alert).

          등급 동률이면 pressure, usage, health, alert 순으로 귀속하며 status 가 unknown (신호 전부

          부재) 이면 생략된다.'
        type: string
      summary:
        type: string
      top_pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.NodePodPressure'
      window:
        type: string
  internal_correlation_api.NodeVitalsResponse:
    type: object
    properties:
      cpu_percent:
        type: number
      generated_at:
        type: string
      gpu_memory_percent:
        type: number
      gpu_memory_total_bytes:
        type: number
      gpu_memory_used_bytes:
        type: number
      gpu_percent:
        type: number
      memory_percent:
        type: number
      node:
        type: string
      summary:
        type: string
  internal_correlation_api.NodesResponse:
    type: object
    properties:
      generated_at:
        type: string
      nodes:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.NodeInventory'
...
  internal_correlation_api.PodCpuDetail:
    type: object
    properties:
      limit_cores:
        type: number
      period_microseconds:
        type: number
      quota_microseconds:
        type: number
      throttled_periods_5m:
        type: number
      throttled_ratio:
        description: ThrottledRatio 는 최근 5분 CFS period 중 throttle 된 비율 (0-1) 이다.
        type: number
      total_periods_5m:
        type: number
  internal_correlation_api.PodDetailResponse:
    type: object
    properties:
      cpu:
        $ref: '#/definitions/internal_correlation_api.PodCpuDetail'
      created_by_kind:
        type: string
      created_by_name:
        type: string
      generated_at:
        type: string
      health:
        description: 'Health 는 차원별 health (0-1, 1 이 healthy) 다. pod 단위 pressure score
          rule 의 1 - score

          환산이며, score 미산출 차원은 엔트리가 생략된다.'
        type: object
        additionalProperties:
          type: number
          format: float64
      namespace:
        type: string
      network:
        $ref: '#/definitions/internal_correlation_api.PodNetworkDetail'
      node:
        type: string
      pod:
        type: string
      pod_ip:
        type: string
      summary:
        type: string
      uid:
        description: 기본 정보는 kube_pod_info 라벨에서 온다. pod 미관측 시 생략된다.
        type: string
      vitals:
        $ref: '#/definitions/internal_correlation_api.PodVitals'
  internal_correlation_api.PodInventory:
    type: object
    properties:
      host_ip:
        type: string
      namespace:
        type: string
      node:
        type: string
      observed:
        description: 'Observed 는 eBPF 관측 커버리지다 (#248). netobs 가 상시 수집하는 netobs_pod_bytes_total

          시리즈가 이 pod 에 존재하면 true, 없으면 no-data 로 판정한다.'
        type: boolean
      phase:
        type: string
      pod:
        type: string
      pod_ip:
        type: string
      priority_class:
        type: string
      qos_class:
        type: string
      uid:
        type: string
      unobserved_reason:
        description: 'UnobservedReason 은 미관측 사유 분류다 (#320, #342). agent_absent (노드에
          netobs 미배치),

          host_network (pod IP 가 host IP 와 같아 IP 귀속이 성립하지 않음), no_traffic (netns 소켓

          테이블이 비어 네트워크 미사용이 증명됨, unix socket 전용 pod 등), no_data (관측 가능한데

          시리즈 부재) 로 구분한다. host_network 는 확정이 아닌 시점 의존 분류로, cgroup 힌트 (#228)

          가 TCP 트래픽에서 학습되면 observed=true (live) 로 전환될 수 있다. observed=true 와

          종료/Unknown phase 는 생략된다.'
        type: string
      workload_kind:
        type: string
      workload_name:
        type: string
  internal_correlation_api.PodMemory:
    type: object
    properties:
      cache_bytes:
        type: number
      dominant_kind:
        type: string
      limit_bytes:
        type: number
      namespace:
        type: string
      node:
        type: string
      oom_risk:
        type: number
      pod:
        type: string
      rss_bytes:
        type: number
      severity:
        type: string
      swap_bytes:
        type: number
      working_set_bytes:
        type: number
  internal_correlation_api.PodNetworkDetail:
    type: object
    properties:
      drop_per_sec:
        type: number
      max_srtt_seconds:
        type: number
      retrans_per_sec:
        type: number
  internal_correlation_api.PodVitals:
    type: object
    properties:
      cpu_percent:
        type: number
      cpu_usage_cores:
        type: number
      memory_percent:
        type: number
      memory_working_set_bytes:
        type: number
  internal_correlation_api.PodsResponse:
    type: object
    properties:
      generated_at:
        type: string
      pods:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.PodInventory'
...
  netobs_internal_apicommon.ErrorBody:
    type: object
    properties:
      error:
        $ref: '#/definitions/netobs_internal_apicommon.ErrorDetail'
  netobs_internal_apicommon.ErrorDetail:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
...